pre-compute sets of files
The gruntfile now pre-computes sets of files that match the rules
currently present in build.py.  The files are exposed in the grunt
configuration object allowing them to be referenced via templates in
other tasks.  This simplifies setting dependencies for tasks and
ensures we are using the same rules as build.py.

A simple grunt task to dump a list by name is added for testing
purposes.  Ideally, it would be modified to execute build.py to dump
the same file list and do a comparison but I don't know how to modify
build.py to do this.  For now it will dump to the console if you run
`grunt listfiles:SRC` and to a file if you run `grunt
listfiles:SRC:build/src.txt`