Adding support for docs
Add support for generating documentation using grunt-jsdoc.  The
documentation task as written in build.py uses the current git branch
for several things.  I've added another grunt task, detect-branch,
which implements this.  Due to the nature of how tasks are configured,
it is not possible to run async code outside of actually executing a
specific task so I settled on this as the cleanest solution.